+++ src/script/util/removeClientFromUserClientMap.ts
